Fix issue that user config won't be used
If you move the configuration file out of home, there is a bit of an
unexpected consequence: `mr register` tries to find a potential
`.mrconfig` and doesn't find any when passing through $HOME.

Since it doesn't find any, it uses... only the built-in configuration.

This commit makes it use the per-user config, if it exists.